# Speed up reading cached results by about 200%